requirements.txt
librosa
soundfile
audioread
webrtcvad
openai-whisper
fugashi[unidic-lite]
resampy
scipy
scikit-learn
numpy
pandas
PyYAML